# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  title: string;
}

export interface OpenaiMessageInput {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export interface DailyProgress {
  /** Weighted learning points accumulated today */
  points: number;
  /** Points needed to reach today's goal */
  target: number;
  /** Progress toward the goal as a percentage (0-100) */
  percent: number;
  /** Whether today's goal has been reached */
  achieved: boolean;
  /** Number of Urdu words/phrases taught today */
  phrasesLearned: number;
}

export interface LessonPhrase {
  /** The phrase written in Urdu script */
  urdu: string;
  /** Roman/English transliteration of the Urdu phrase */
  transliteration: string;
  /** English meaning */
  english: string;
}

export interface Lesson {
  slug: string;
  title: string;
  description: string;
  /** A short label or icon hint for the lesson topic */
  emoji?: string;
  phrases: LessonPhrase[];
}