# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: openai
    description: OpenAI AI chat operations
  - name: lessons
    description: Urdu lesson content
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"
  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiMessageInput"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}
  /progress/today:
    get:
      operationId: getProgressToday
      tags: [progress]
      summary: Today's learning progress
      description: Returns the student's progress toward today's learning goal.
      responses:
        "200":
          description: Today's progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyProgress"
  /lessons:
    get:
      operationId: listLessons
      tags: [lessons]
      summary: List beginner Urdu lessons
      description: Returns structured beginner Urdu lessons with vocabulary and phrases.
      responses:
        "200":
          description: List of lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"
  /lessons/{slug}:
    get:
      operationId: getLesson
      tags: [lessons]
      summary: Get a single lesson by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lesson detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    OpenaiConversationInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
      required:
        - title
    OpenaiMessageInput:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 4000
      required:
        - content
    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages
    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    DailyProgress:
      type: object
      properties:
        points:
          type: integer
          description: Weighted learning points accumulated today
        target:
          type: integer
          description: Points needed to reach today's goal
        percent:
          type: integer
          description: Progress toward the goal as a percentage (0-100)
        achieved:
          type: boolean
          description: Whether today's goal has been reached
        phrasesLearned:
          type: integer
          description: Number of Urdu words/phrases taught today
      required:
        - points
        - target
        - percent
        - achieved
        - phrasesLearned
    LessonPhrase:
      type: object
      properties:
        urdu:
          type: string
          description: The phrase written in Urdu script
        transliteration:
          type: string
          description: Roman/English transliteration of the Urdu phrase
        english:
          type: string
          description: English meaning
      required:
        - urdu
        - transliteration
        - english
    Lesson:
      type: object
      properties:
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        emoji:
          type: string
          description: A short label or icon hint for the lesson topic
        phrases:
          type: array
          items:
            $ref: "#/components/schemas/LessonPhrase"
      required:
        - slug
        - title
        - description
        - phrases