# lib/api-zod/src/generated/types/dailyProgress.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DailyProgress {
  /** Weighted learning points accumulated today */
  points: number;
  /** Points needed to reach today's goal */
  target: number;
  /** Progress toward the goal as a percentage (0-100) */
  percent: number;
  /** Whether today's goal has been reached */
  achieved: boolean;
  /** Number of Urdu words/phrases taught today */
  phrasesLearned: number;
}